Added updated humanoid file
This xml file is an edited version of the DeepMind humanoid, with a balance board added for the humanoid to stand on. The feet of the humanoid are welded to the board due to the suboptimal friction of MuJoCo simulations. The mesh for the board can be found in the resources folder.